+++ components/AppBar/examples/AppBarTypicalUseExample.m
@@ -48,7 +48,7 @@
 
     [self addChildViewController:_appBarViewController];
 
-    _appBarViewController.navigationBar.inkColor = [UIColor colorWithWhite:0.9f alpha:0.1f];
+    _appBarViewController.navigationBar.inkColor = [UIColor colorWithWhite:(CGFloat)0.9 alpha:(CGFloat)0.1];
 
     self.colorScheme = [[MDCSemanticColorScheme alloc] init];
     self.typographyScheme = [[MDCTypographyScheme alloc] init];
